notifications kind of work
All the text is finished for the notifications, but no email is actually sent. Instead, the email function just takes the recipient, subject, and body, and prints this out to console.

Worse yet: while the server fires notifications (as fake emails) for events that exist before the server started, it somehow fails to catch them if they fire while the server is running. I currently have some debug messages to help identify the problem, but no luck so far...

This will be my last commit for a while.